Remove Negative margin to adjust the UI
Edit Message Action buttons were getting cropped due to this.
