+++ src/libs/reportUtils.js
@@ -2,6 +2,7 @@ import _ from 'underscore';
 import Str from 'expensify-common/lib/str';
 import lodashGet from 'lodash/get';
 import Onyx from 'react-native-onyx';
+import moment from 'moment';
 import ONYXKEYS from '../ONYXKEYS';
 import CONST, {EXPENSIFY_EMAILS} from '../CONST';
 
@@ -167,6 +168,22 @@ function hasExpensifyEmails(emails) {
     return _.intersection(emails, EXPENSIFY_EMAILS).length > 0;
 }
 
+function shouldShowReportRecipientLocalTime(personalDetails, myPersonalDetails, report) {
+    // eslint-disable-next-line no-unused-vars
+    const reportParticipants = lodashGet(report, 'participants', []);
+    const hasMultipleParticipants = reportParticipants.length > 1;
+    const reportRecipient = personalDetails[reportParticipants[0]];
+    const currentUserTimezone = lodashGet(myPersonalDetails, 'timezone', CONST.DEFAULT_TIME_ZONE);
+    const reportRecipientTimezone = lodashGet(reportRecipient, 'timezone', CONST.DEFAULT_TIME_ZONE);
+    return !hasExpensifyEmails(reportParticipants)
+        && !hasMultipleParticipants
+        && reportRecipient
+        && reportRecipientTimezone
+        && currentUserTimezone.selected
+        && reportRecipientTimezone.selected
+        && moment().tz(currentUserTimezone.selected).utcOffset() !== moment().tz(reportRecipientTimezone.selected).utcOffset();
+}
+
 export {
     getReportParticipantsTitle,
     isReportMessageAttachment,
@@ -179,4 +196,5 @@ export {
     isArchivedRoom,
     isConciergeChatReport,
     hasExpensifyEmails,
+    shouldShowReportRecipientLocalTime,
 };
